new try

+++ hashmap.c
@@ -70,7 +70,7 @@ void insertMap(HashMap * map, char * key, void * value) {
 // Recuerde actualizar el índice current a la posición encontrada. Recuerde que el arreglo es circular.
 
 Pair * searchMap(HashMap * map,  char * key) {
-    int pos = key % map->size;
+    long pos = key % map->size;
     if(is_equal(map->buckets[pos]->key,key) == 1)return map->buckets[pos]->value;
     while(map->buckets[pos]->key != NULL){
         pos = (pos +1)% map->capacity;
